refactor agentskills to prepare for agentless
