+++ opendevin/runtime/plugins/agent_skills/file_reader/__init__.py
@@ -1,0 +1,4 @@
+from ..utils.modules import import_modules
+from .file_readers import __all__
+
+import_modules(__name__ + '.file_readers', __all__)
